backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
python-docx==1.1.0
PyPDF2==3.0.1
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.0
